pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gmcore-dashboard"
version = "0.1.0"
description = "GMCORE experiment management and visualization dashboard"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "dash>=2.0",
    "dash-bootstrap-components>=1.0",
    "dash-ace>=0.2",
    "plotly>=5.0",
    "numpy>=1.21",
    "xarray>=2023.1.0",
    "netCDF4>=1.6.4",
    "diskcache",
    "f90nml>=1.4",
    "pyyaml>=6.0",
    "markdown",
    "flask",
    "werkzeug",
]

[project.scripts]
gmcore-dashboard = "gmcore_dashboard.cli:main"
gmcore-experiments = "gmcore_dashboard.experiments.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["gmcore_dashboard"]
